perf(waf): unify hot-path to Aho-Corasick O(N) and scale connection pool
Replace O(N*M) linear scan in detect_waf_challenge with Aho-Corasick
O(N) via new WafInspector::detect_body() method. Zero API change for
callers — same Option<&'static str> return type.

Changes:
- Add detect_body() with AC automaton, 62 merged signatures, Shannon
  entropy detection for >100KB obfuscated bodies
- Swap 3 callers (scraper_service, client, mcp_server) to detect_body()
- Delete waf.rs (304 lines of dead O(N*M) code)
- Increase pool_max_idle_per_host from 3 to 6 for HDD socket reuse
- Port 22 unit tests + update 8 security integration test call sites

Tested: 669/669 passed, 0 clippy warnings
