+++ src/infrastructure/http/waf_engine.rs
@@ -4,16 +4,22 @@
 //! in http_client.rs. It includes:
 //! - Detection by Control Headers (x-datadome-response, cf-mitigated, etc.)
 //! - Entropy analysis for "Silent Challenge" detection
-//! - Efficient O(N) matching using Aho-Corasick for 50+ signatures
+//! - Efficient O(N) matching using Aho-Corasick for 60+ signatures
+//! - Body-only detection via `detect_body()` for callers that only have the body
 //!
 //! # Usage
 //!
 //! ```rust
 //! use rust_scraper::infrastructure::http::waf_engine::WafInspector;
 //!
-//! // After receiving HTTP 200 response with body
+//! // Full integrity check (headers + body)
 //! if let Err(e) = WafInspector::verify_integrity(&response, &body) {
-//!     return Err(e); // WAF challenge detected
+//!     return Err(e);
+//! }
+//!
+//! // Body-only check (replaces legacy detect_waf_challenge)
+//! if let Some(provider) = WafInspector::detect_body(&body) {
+//!     eprintln!("WAF detected: {provider}");
 //! }
 //! ```
 
@@ -34,61 +40,91 @@ const WAF_CONTROL_HEADERS: &[(&str, &str)] = &[
     ("x-cdn", "Imperva"),
 ];
 
-/// WAF signatures for HTML body scanning (2026 updated)
-/// These replace the basic signatures in http_client.rs for more comprehensive coverage
+/// Merged WAF signatures for body scanning (62 unique patterns).
+/// Combined from legacy `waf.rs` (41) + `waf_engine.rs` (39), deduplicated by pattern string.
 const WAF_BODY_SIGNATURES: &[(&str, &str)] = &[
-    // Cloudflare (updated 2026)
+    // Cloudflare
     ("cf-turnstile", "Cloudflare Turnstile"),
     ("challenge-platform", "Cloudflare JS Challenge"),
     ("Just a moment...", "Cloudflare"),
     ("Checking your browser", "Cloudflare"),
     ("__cf_chl_f_tk", "Cloudflare"),
+    ("cf-browser-verification", "Cloudflare"),
+    ("cf-ray", "Cloudflare"),
+    ("cf-cache-status", "Cloudflare"),
     ("_cf_chl_opt", "Cloudflare"),
     ("cloudflare", "Cloudflare"),
     ("cf-dns", "Cloudflare"),
-    // Google reCAPTCHA (updated)
+    // Google reCAPTCHA
     ("g-recaptcha", "reCAPTCHA"),
     ("recaptcha/api.js", "reCAPTCHA"),
     ("grecaptcha.execute", "reCAPTCHA"),
+    ("recaptcha.net", "reCAPTCHA"),
     ("recaptcha Enterprise", "reCAPTCHA"),
-    // hCaptcha (updated)
+    // hCaptcha
     ("hcaptcha.com", "hCaptcha"),
     ("h-captcha", "hCaptcha"),
+    ("hcaptcha-api", "hCaptcha"),
     ("hcaptcha.js", "hCaptcha"),
-    // DataDome (updated 2026 - Silent Challenges)
+    // DataDome
     ("datadome", "DataDome"),
     ("dd-captcha", "DataDome"),
+    ("datadome.co", "DataDome"),
     ("dd=", "DataDome"),
     ("data-domain", "DataDome"),
-    // PerimeterX / HUMAN Security (updated)
+    // PerimeterX / HUMAN Security
     ("perimeterx", "PerimeterX"),
     ("_pxCaptcha", "PerimeterX"),
+    ("px-captcha", "PerimeterX"),
+    ("perimeterx.net", "PerimeterX"),
     ("human-security", "HUMAN"),
     ("px-init", "PerimeterX"),
-    // Akamai Bot Manager (updated 2026)
+    // Akamai Bot Manager
     ("_abck", "Akamai Bot Manager"),
     ("SensorData", "Akamai Bot Manager"),
+    ("akamai-bot-manager", "Akamai Bot Manager"),
+    ("akamai.net", "Akamai"),
     ("akamai", "Akamai"),
-    // Imperva (updated)
+    // Imperva / Incapsula
     ("imperva", "Imperva"),
     ("incapsula", "Imperva"),
     ("_Incapsula_Resource", "Imperva"),
-    // F5 (updated)
+    ("visid_incap", "Imperva Incapsula"),
+    ("incap_ses", "Imperva Incapsula"),
+    // Sucuri
+    ("sucuri", "Sucuri"),
+    ("sucuri.net", "Sucuri"),
+    // F5
     ("_nfv", "F5"),
     ("BIGipServer", "F5"),
-    // Generic challenges (expanded)
+    // Generic challenges
     ("Please verify you are a human", "Generic Challenge"),
     ("verify you are human", "Generic Challenge"),
     ("bot detection", "Generic Detection"),
+    ("automated requests", "Generic Detection"),
+    ("security check", "Generic Challenge"),
+    ("anti-bot", "Generic Detection"),
     ("checking your browser", "Browser Verification"),
     ("attack detected", "Security Firewall"),
     ("suspicious activity", "Security Firewall"),
     ("captcha-delivery", "Challenge Delivery"),
     ("__js_challenge__", "JS Challenge"),
+    // Generic JS/CAPTCHA scripts
+    ("challenge.js", "Generic Challenge"),
+    ("captcha.js", "Generic Challenge"),
+    ("verify.js", "Generic Challenge"),
+    ("bot-check", "Generic Detection"),
 ];
 
-/// Aho-Corasick automaton for O(N) multi-pattern matching
-/// Replaces O(N*M) linear search in http_client.rs
+/// Shannon entropy threshold for obfuscated WAF detection
+const ENTROPY_THRESHOLD: f64 = 5.5;
+
+/// Body size threshold (100KB) above which entropy analysis is applied
+const SUSPICIOUS_SIZE_THRESHOLD: usize = 100_000;
+
+/// Aho-Corasick automaton for O(N) multi-pattern body matching
+/// Replaces O(N*M) linear scan in legacy `waf.rs`.
+/// Compiled once via `Lazy`, thread-safe for concurrent reads.
 static WAF_AC: Lazy<AhoCorasick> = Lazy::new(|| {
     AhoCorasick::new(WAF_BODY_SIGNATURES.iter().map(|(sig, _)| sig))
         .expect("Failed to build Aho-Corasick automaton")
@@ -98,6 +134,40 @@ static WAF_AC: Lazy<AhoCorasick> = Lazy::new(|| {
 pub struct WafInspector;
 
 impl WafInspector {
+    /// Scan body for WAF challenge signatures using Aho-Corasick (O(N) single pass).
+    ///
+    /// Returns the FIRST matching provider name, or `None` if the body is clean.
+    /// For bodies exceeding 100KB, Shannon entropy is computed; if entropy > 5.5,
+    /// returns `Some("Obfuscated WAF")`.
+    ///
+    /// Thread-safe: the AC automaton is immutable once compiled via `Lazy`.
+    ///
+    /// # Arguments
+    /// * `body` - The HTTP response body to scan
+    ///
+    /// # Returns
+    /// * `Some(provider_name)` - WAF challenge detected
+    /// * `None` - No WAF challenge detected
+    #[must_use]
+    pub fn detect_body(body: &str) -> Option<&'static str> {
+        // Early exit for empty or very small bodies (no signatures fit in <10 chars)
+        if body.len() < 10 {
+            return None;
+        }
+
+        // Shannon entropy check for large bodies (>100KB)
+        if body.len() > SUSPICIOUS_SIZE_THRESHOLD {
+            let entropy = calculate_entropy(body);
+            if entropy > ENTROPY_THRESHOLD {
+                return Some("Obfuscated WAF");
+            }
+        }
+
+        // Aho-Corasick single-pass scan for all 62 patterns.
+        // Returns provider name for the first match found by AC (earliest end position).
+        WAF_AC.find(body).map(|m| WAF_BODY_SIGNATURES[m.pattern()].1)
+    }
+
     /// Verify response integrity across multiple layers
     ///
     /// 1. Control Headers: Check for WAF-specific headers (immediate)
@@ -224,9 +294,186 @@ impl WafInspector {
     }
 }
 
+/// Calculate Shannon entropy of a string
+///
+/// Used to detect obfuscated JavaScript in challenge pages, which often have
+/// high entropy due to minification and encoding.
+///
+/// # Arguments
+/// * `s` - The string to analyze
+///
+/// # Returns
+/// * Entropy value between 0.0 (uniform distribution) and ~8.0 (high entropy)
+#[inline]
+fn calculate_entropy(s: &str) -> f64 {
+    if s.is_empty() {
+        return 0.0;
+    }
+
+    let mut freq = [0u32; 256];
+    let len = s.len() as f64;
+
+    for &b in s.as_bytes() {
+        freq[b as usize] += 1;
+    }
+
+    let mut entropy = 0.0;
+    for &count in &freq {
+        if count > 0 {
+            let p = count as f64 / len;
+            entropy -= p * p.log2();
+        }
+    }
+
+    entropy
+}
+
+/// Check if body size indicates a potential WAF challenge (>100KB)
+#[cfg(test)]
+#[inline]
+fn is_suspicious_size(body_len: usize) -> bool {
+    body_len > SUSPICIOUS_SIZE_THRESHOLD
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    // ========================================================================
+    // detect_body() tests — ported from waf.rs (Approval Testing)
+    // ========================================================================
+
+    #[test]
+    fn test_detect_body_cloudflare_turnstile() {
+        let html = r#"<div id="cf-turnstile" data-sitekey="abc123"></div>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("Cloudflare Turnstile"));
+    }
+
+    #[test]
+    fn test_detect_body_cloudflare_just_a_moment() {
+        let html = "<html><body><h1>Just a moment...</h1></body></html>";
+        assert_eq!(WafInspector::detect_body(html), Some("Cloudflare"));
+    }
+
+    #[test]
+    fn test_detect_body_cloudflare_checking_browser() {
+        let html = "<html><body>Checking your browser before accessing...</body></html>";
+        assert_eq!(WafInspector::detect_body(html), Some("Cloudflare"));
+    }
+
+    #[test]
+    fn test_detect_body_recaptcha() {
+        let html = r#"<script src="https://www.google.com/recaptcha/api.js?render=abc"></script>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("reCAPTCHA"));
+    }
+
+    #[test]
+    fn test_detect_body_g_recaptcha() {
+        let html = r#"<div class="g-recaptcha" data-sitekey="abc"></div>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("reCAPTCHA"));
+    }
+
+    #[test]
+    fn test_detect_body_hcaptcha() {
+        let html = r#"<div class="h-captcha" data-sitekey="abc"></div>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("hCaptcha"));
+    }
+
+    #[test]
+    fn test_detect_body_datadome() {
+        let html = r#"<script src="https://js.datadome.co/captcha.js"></script>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("DataDome"));
+    }
+
+    #[test]
+    fn test_detect_body_perimeterx() {
+        let html = r#"<script>var _pxCaptcha = {};</script>"#;
+        assert_eq!(WafInspector::detect_body(html), Some("PerimeterX"));
+    }
+
+    #[test]
+    fn test_detect_body_akamai() {
+        let html = r#"<input type="hidden" name="_abck" value="xxx">"#;
+        assert_eq!(WafInspector::detect_body(html), Some("Akamai Bot Manager"));
+    }
+
+    #[test]
+    fn test_detect_body_generic_challenge() {
+        let html = "<p>Please verify you are a human to continue.</p>";
+        assert_eq!(WafInspector::detect_body(html), Some("Generic Challenge"));
+    }
+
+    #[test]
+    fn test_detect_body_clean_html() {
+        let html = r#"
+            <html>
+                <head><title>Normal Page</title></head>
+                <body>
+                    <article>
+                        <h1>Welcome</h1>
+                        <p>This is a normal page with real content.</p>
+                    </article>
+                </body>
+            </html>
+        "#;
+        assert_eq!(WafInspector::detect_body(html), None);
+    }
+
+    #[test]
+    fn test_detect_body_empty() {
+        assert_eq!(WafInspector::detect_body(""), None);
+    }
+
+    // ========================================================================
+    // Entropy tests — ported from waf.rs
+    // ========================================================================
+
+    #[test]
+    fn test_calculate_entropy_high() {
+        let obfuscated_js: String = (0u8..=255).map(|b| b as char).collect();
+        let entropy = calculate_entropy(&obfuscated_js);
+        assert!(entropy > 6.0, "entropy={entropy}, expected > 6.0");
+    }
+
+    #[test]
+    fn test_calculate_entropy_low() {
+        let plain_text = "Hello world, this is a normal page with regular content.";
+        let entropy = calculate_entropy(plain_text);
+        assert!(entropy < 5.0);
+    }
+
+    #[test]
+    fn test_is_suspicious_size() {
+        assert!(is_suspicious_size(150_000));
+        assert!(!is_suspicious_size(10_000));
+        assert!(!is_suspicious_size(100_000));
+        assert!(is_suspicious_size(100_001));
+    }
+
+    #[test]
+    fn test_detect_body_by_entropy() {
+        // Create >100KB with high entropy to trigger Shannon entropy detection
+        let high_entropy_content: String = (0u8..=255)
+            .map(|b| b as char)
+            .chain((0u8..=255).map(|b| b as char))
+            .chain((0u8..=255).map(|b| b as char))
+            .chain((0u8..=255).map(|b| b as char))
+            .cycle()
+            .take(104_000)
+            .collect();
+        let result = WafInspector::detect_body(&high_entropy_content);
+        assert_eq!(result, Some("Obfuscated WAF"));
+    }
+
+    #[test]
+    fn test_detect_body_small_low_entropy() {
+        let small_content = "<html><body>Redirecting...</body></html>";
+        assert_eq!(WafInspector::detect_body(small_content), None);
+    }
+
+    // ========================================================================
+    // verify_integrity() tests (existing, unchanged)
+    // ========================================================================
 
     #[test]
     fn test_waf_control_header_detection() {
@@ -265,12 +512,10 @@ mod tests {
 
     #[test]
     fn test_silent_challenge_detection() {
-        // High script density in small body should trigger
         let body = r#"<html><script></script><script></script><script></script><script></script><script></script><script></script></html>"#;
         let result = WafInspector::verify_integrity(&HeaderMap::new(), body);
         assert!(result.is_err());
 
-        // Normal content should pass
         let body = "<html><body><p>Hello</p></body></html>";
         let result = WafInspector::verify_integrity(&HeaderMap::new(), body);
         assert!(result.is_ok());
@@ -278,7 +523,6 @@ mod tests {
 
     #[test]
     fn test_aho_corasick_performance() {
-        // Test that Aho-Corasick works correctly - using exact phrase from signatures
         let body = "This is a page with Just a moment... and recaptcha/api.js content";
         let result = WafInspector::verify_integrity(&HeaderMap::new(), body);
         assert!(result.is_err());
